# Optimize zip using flate2