--- a/Cargo.toml
+++ b/Cargo.toml
@@ -24,7 +24,7 @@ serde-xml-rs = "0.3.1"
 subprocess = "0.1"
 toml = "0.5"
 walkdir = "2.2"
-zip = { version = "0.5.2", features = ["deflate"] }
+zip = { git = "https://github.com/jonpas/zip-rs", branch = "flate2", features = ["deflate"] }
 
 [target.'cfg(windows)'.dependencies]
 ansi_term = "0.11"