**When merged this pull request will:**
- Replace #133 
- Optimize `zip` utility (better than #133)
  - Using `flate2` (DE)FLATE implementation in `zip-rs` from [`jonpas/zip-rs [flate2]`](https://github.com/jonpas/zip-rs/tree/flate2) instead `libflate`
  - Ref. https://github.com/mvdnes/zip-rs/issues/88#issuecomment-491185791 and https://github.com/mvdnes/zip-rs/pull/107

In debug build, compared to #133, shaved off another ~30%. `release` build compression speeds are now on-par with 7z on Windows.

## Benchmarks

All benchmarks were ran on `release` build (`cargo build --release`).
- **Old**: HEMTT 249d6e8a37a3390a798bd7ad98bf9313b1dc2297
- **#133**: Pull request #133 (`BufReader`/`BufWriter`/`copy` optimization)
- **New**: **#133** + `flake2` implementation in `zip-rs` (instead of `libflate`)

_Difference in **ACE3** Folder Size is due to binarization._
_**Old** and **#133** Zip Size is the same as (DE)FLATE implementation is the same._

### Linux

Mod | Folder Size | Old Speed | #133 Speed | New Speed | Old / #133 Zip Size | New Zip Size
-------- | ------- | ----- | ----- | ----- | ------- | -------
**TACU** | 12.2 MB | 9 s   | 1.5 s | 0.5 s | 11.2 MB | 11.2 MB
**TACR** | 59.2 MB | 44 s  | 7.5 s | 2.5 s | 50.6 MB | 50.5 MB
**ACE3** | 205 MB  | 110 s | 15 s  | 7.2 s | 143 MB  | 139 MB

### Windows

Mod | Folder Size | Old Speed | #133 Speed | New Speed | Old / #133 Zip Size | New Zip Size
-------- | ------- | ----- | ----- | ----- | ------- | -------
**TACU** | 13 MB   | 13 s  | 2.5 s | 0.5 s | 11.2 MB | 11.1 MB
**TACR** | 59.2 MB | 65 s  | 12 s | 3 s | 50.6 MB | 50.4 MB
**ACE3** | 174 MB  | 150 s | 30 s  | 8.5 s | 137 MB  | 135 MB

### `rust_backend` ([`miniz_oxide`](https://crates.io/crates/miniz_oxide))

No noticeable difference compared to default `miniz`.